Next.js v16 -> v15 to enable use of NextAuth dependencies
