+++ package.json
@@ -21,8 +21,8 @@
     "firebase": "^10.12.3",
     "firebase-admin": "^13.5.0",
     "googleapis": "^164.0.0",
-    "next": "^16.0.0",
-    "next-auth": "4.24.7",
+    "next": "^15.5.6",
+    "next-auth": "^4.24.11",
     "nodemailer": "^6.9.14",
     "pizzip": "^3.2.0",
     "react": "^18",
